agents/data-ingestion-agent/requirements.txt
# Core dependencies
llama-index>=0.10.0
llama-index-core>=0.10.0

# CLI and utilities
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
pydantic>=2.0.0

# File processing
pypdf>=3.17.0

# Web scraping (optional)
beautifulsoup4>=4.12.0
requests>=2.31.0
llama-index-readers-web>=0.1.0

# Database support (optional)
sqlalchemy>=2.0.0
llama-index-readers-database>=0.1.0

# Additional readers
llama-index-readers-file>=0.1.0
